pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "pydemo"
version = "0.1.0"
description = "A Python project for stock data processing"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "requests",
    "confluent-kafka",
    "python-dotenv",
    "click",
]